implements some cart controllers

+++ api/src/modules/carts/cart.controller.ts
@@ -26,11 +26,17 @@ export const getCartByUser = async (req: Request, res: Response) => {
 export const updateCart = async (req: Request, res: Response) => {
   const { cartId } = req.params;
   const { body } = req;
+
   try {
     let cart = await CartService.getCart(cartId);
     if (!cart) {
       return res.status(404).json({ message: "Cart not found" });
     }
+
+    if (cart.userId !== req.user?.id) {
+      return res.status(403).json({ message: "You are not permitted" });
+    }
+
     cart = await CartService.updateCart(cartId, body);
     res.status(200).json({ cart });
   } catch (error: any) {
@@ -41,7 +47,14 @@ export const updateCart = async (req: Request, res: Response) => {
 export const deleteCart = async (req: Request, res: Response) => {
   const { cartId } = req.params;
   try {
-    CartService.deleteCart(cartId);
+    let cart = await CartService.deleteCart(cartId);
+
+    if (cart && cart.id !== req.user?.id) {
+      return res.status(403).json({ message: "Forbidden" });
+    }
+
+    await CartService.deleteCart(cartId);
+
     res.status(200).json({ message: "Cart Delete successfully" });
   } catch (error: any) {
     return res.status(500).json(error.message);
